Moves tests for download_pdfs() to separate file
